Move DryRunVerifier from kubectl to cli-runtime
