+++ staging/src/k8s.io/cli-runtime/pkg/resource/BUILD
@@ -7,6 +7,7 @@ go_library(
         "client.go",
         "crd_finder.go",
         "doc.go",
+        "dry_run_verifier.go",
         "fake.go",
         "helper.go",
         "interfaces.go",
@@ -44,6 +45,7 @@ go_library(
         "//staging/src/k8s.io/client-go/kubernetes/scheme:go_default_library",
         "//staging/src/k8s.io/client-go/rest:go_default_library",
         "//staging/src/k8s.io/client-go/restmapper:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/util/openapi:go_default_library",
         "//vendor/golang.org/x/text/encoding/unicode:go_default_library",
         "//vendor/golang.org/x/text/transform:go_default_library",
         "//vendor/sigs.k8s.io/kustomize/pkg/fs:go_default_library",
@@ -55,6 +57,7 @@ go_test(
     srcs = [
         "builder_test.go",
         "crd_finder_test.go",
+        "dry_run_verifier_test.go",
         "helper_test.go",
         "scheme_test.go",
         "visitor_test.go",
@@ -84,7 +87,9 @@ go_test(
         "//staging/src/k8s.io/client-go/restmapper:go_default_library",
         "//staging/src/k8s.io/client-go/util/testing:go_default_library",
         "//vendor/github.com/davecgh/go-spew/spew:go_default_library",
+        "//vendor/github.com/googleapis/gnostic/OpenAPIv2:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
+        "//vendor/k8s.io/kube-openapi/pkg/util/proto/testing:go_default_library",
         "//vendor/sigs.k8s.io/yaml:go_default_library",
     ],
 )
