+++ staging/src/k8s.io/kubectl/pkg/cmd/annotate/annotate.go
@@ -55,7 +55,7 @@ type AnnotateOptions struct {
 	local           bool
 	dryRunClient    bool
 	dryRunServer    bool
-	dryRunVerifier  *cmdutil.DryRunVerifier
+	dryRunVerifier  *resource.DryRunVerifier
 	all             bool
 	resourceVersion string
 	selector        string
@@ -181,7 +181,7 @@ func (o *AnnotateOptions) Complete(f cmdutil.Factory, cmd *cobra.Command, args [
 	if err != nil {
 		return err
 	}
-	o.dryRunVerifier = &cmdutil.DryRunVerifier{
+	o.dryRunVerifier = &resource.DryRunVerifier{
 		Finder:        resource.NewCRDFinder(resource.CRDFromDynamic(dynamicClient)),
 		OpenAPIGetter: discoveryClient,
 	}
